Folder name-HtmlPrac
This Code is HTML only. 

The purpose is to study when and where to use the most effective HTML tags and empty elements
across websites.

I am practicing using the HTML tags and empty elements.
I am learning when and where to apply the tags in text.
I am learning how to style the actual ID to collaborate with other code writers ie: the White space and White space collapsing technique.